Update support Debian Versions
This removes support for Wheezy and Jesse and confirms support for
Stretch and Buster.

Defaults that applied only to Stretch have been made standard.

Stretch explicitly installed ssl-cert, as this is not a formal dep but
is needed by this config. Note that is has become a dep in Buster.