Use Enum type for some SSL module parameters
The value of `ssl` can be only one of "yes", "no" or "required" so we
can specify this in the paramter declaration.

This also enforces "TLSv1.2" as the only currently acceptable protocol
string. Rather than hardcoding this, we'll set it this way to prevent
the use of earlier protocols that are now deprecated.

Dovecot 2.3.15 will add support for TLSv1.3 and also "LATEST", but this
isn't currently available in the core Debian repositories for stable
distributions (2.3.19 is currently in testing and unstable).